refactor: extract hotkey fleet actions
